+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Fixed #87 Dataprocessors with required contact id filters throws "One of parameters (value: ) is not of the type Int" error for anonymous users by !77
 * Compatibility fix for Symfony 2.8, 3.4 and 4.0
+* Fixed issue with non required join and filtering on case type.
 
 # Version 1.34